+++ src/assets/styles/FontSizes.json
@@ -30,7 +30,7 @@
     "QZ_FORM_INP": "1.6rem",
     "QZ_FORM_LB_CB": "0.8rem",
     "QZ_FORM_BTN": "0.8rem",
-    "QN_BTN": "0.8rem",
-    "QN_COUNT": "1rem"
+    "QN_NAV_BTN": "0.8rem",
+    "QN_NAV_COUNT": "1rem"
   }
 }
