--- src/pages/quiz/QnAzContainer.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function QnAzContainer() {
-  return <div></div>;
-}
